revert type change for rpt_period_id changes

+++ src/entities/monitor-plan.entity.ts
@@ -64,14 +64,14 @@ export class MonitorPlan extends BaseEntity {
 
   @Column({
     name: 'begin_rpt_period_id',
-    type: 'bigint',
+    type: 'numeric',
     transformer: new NumericColumnTransformer(),
   })
   beginRPTPeriodIdentifier: number;
 
   @Column({
     name: 'end_rpt_period_id',
-    type: 'bigint',
+    type: 'numeric',
     transformer: new NumericColumnTransformer(),
   })
   endRPTPeriodIdentifier: number;
