+++ src/entities/test-summary.entity.ts
@@ -47,7 +47,7 @@ export class TestSummary extends BaseEntity {
 
   @Column({
     name: 'rpt_period_id',
-    type: 'bigint',
+    type: 'numeric',
     transformer: new NumericColumnTransformer(),
   })
   rptPeriodIdentifier: number;
